requirements.txt
alembic==1.15.1
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
mysqlclient==2.2.7
packaging==24.2
paypalrestsdk==1.13.3
pycparser==2.22
pyOpenSSL==25.0.0
python-dotenv==1.0.1
python-engineio==4.11.2
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.39
stripe==11.6.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
